♻️ Extract duplicated utility functions to centralized lib
- Added extractTitleFromContent() to lib/utils.ts
- Added getWordCount() to lib/utils.ts
- Removed duplicate extractTitleFromContent from use-note-management hook
- Removed duplicate extractTitleFromContent from DetachedNoteWindow
- Replaced all word count calculations with centralized getWordCount()
- Improved code maintainability and consistency
